Update to use a single build name for the lifetime of a test suite run
