+++ UI.NUnitVersion/Drivers/RemoteChromeVhDriver.cs
@@ -1,18 +1,17 @@
-using System.Globalization;
 using OpenQA.Selenium.Remote;
+using UI.NUnitVersion.Utilities;
 
 namespace UI.NUnitVersion.Drivers;
 
 public class RemoteChromeVhDriver : IVhDriver
 {
     private RemoteWebDriver _driver;
-    private readonly ChromeOptions _driverOptions;
 
     public RemoteChromeVhDriver(string platform = "Windows 11", string browserVersion = "latest")
     {
         var envConfigSettings = ConfigRootBuilder.Build().GetSection("SystemConfiguration:EnvironmentConfigSettings")
             .Get<EnvironmentConfigSettings>();
-        _driverOptions = new ChromeOptions
+        var driverOptions = new ChromeOptions
         {
             PlatformName = platform,
             BrowserVersion = browserVersion
@@ -20,12 +19,12 @@ public class RemoteChromeVhDriver : IVhDriver
 
         var buildName = Environment.GetEnvironmentVariable("TF_BUILD") == null ? 
             $"local-{Environment.MachineName}-{Environment.UserName}-{DateTime.Now:dd-mm-yy-hh-mm}" : 
-            GetBuildNameForSauceLabs();
+            BuildName.GetBuildNameForSauceLabs(driverOptions.BrowserName, driverOptions.BrowserVersion, driverOptions.PlatformName);
 
         var sauceLabsConfiguration = envConfigSettings.SauceLabsConfiguration;
         
-        _driverOptions.AddAdditionalOption("username", sauceLabsConfiguration.SauceUsername);
-        _driverOptions.AddAdditionalOption("accessKey", sauceLabsConfiguration.SauceAccessKey);
+        driverOptions.AddAdditionalOption("username", sauceLabsConfiguration.SauceUsername);
+        driverOptions.AddAdditionalOption("accessKey", sauceLabsConfiguration.SauceAccessKey);
         
         var sauceOptions = new Dictionary<string, object>
         {
@@ -37,10 +36,10 @@ public class RemoteChromeVhDriver : IVhDriver
             {"idleTimeout", sauceLabsConfiguration.IdleTimeoutInSeconds},
             {"screenResolution", sauceLabsConfiguration.WindowsScreenResolution},
         };
-        _driverOptions.AddAdditionalOption("sauce:options", sauceOptions);
+        driverOptions.AddAdditionalOption("sauce:options", sauceOptions);
         
         var remoteUrl = new Uri($"https://{sauceLabsConfiguration.SauceUsername}:{sauceLabsConfiguration.SauceAccessKey}@{sauceLabsConfiguration.SecureSauceUrl}");
-        var remoteDriver = new RemoteWebDriver(remoteUrl, _driverOptions);
+        var remoteDriver = new RemoteWebDriver(remoteUrl, driverOptions);
         remoteDriver.FileDetector = new LocalFileDetector();
         _driver = remoteDriver;
     }
@@ -78,36 +77,5 @@ public class RemoteChromeVhDriver : IVhDriver
         }
     }
     
-    private string GetBuildNameForSauceLabs()
-    {
-        var attemptNumber = GetAttemptNumber();
-        var build = $"{GetBuildDefinition()}{GetGitVersionNumber()} {DateTime.Now:dd-mm-yy-hh-mm}     [ {_driverOptions.BrowserName} | {_driverOptions.PlatformName} | {_driverOptions.BrowserVersion} ] {attemptNumber}";
-        return build;
-    }
     
-    private static string GetAttemptNumber()
-    {
-        var attemptNumber = Environment.GetEnvironmentVariable("Build_AttemptNumber");
-        if (string.IsNullOrWhiteSpace(attemptNumber)) return string.Empty;
-        return Convert.ToInt32(attemptNumber) > 1 ? $" : Attempt {attemptNumber}" : string.Empty;
-    }
-        
-    private static string GetGitVersionNumber()
-    {
-        var gitVersionNumber = Environment.GetEnvironmentVariable("GITVERSION_FULLSEMVER");
-        return !string.IsNullOrEmpty(gitVersionNumber) ? $" | {gitVersionNumber}" : string.Empty;
-    }
-        
-    private static string GetBuildDefinition()
-    {
-        var definition = Environment.GetEnvironmentVariable("BUILD_DEFINITIONNAME")?
-                             .ToLower()
-                             .Replace("hmcts.vh-", "")
-                             .Replace("-", " ")
-                             .Replace("cd", "")
-                             .Replace("webnightly", " Web Nightly")
-                             .Replace("web", " Web")
-                         ?? string.Empty;
-        return new CultureInfo("en-GB", false).TextInfo.ToTitleCase(definition);
-    }
 }
